Fixed bug in dassl.jacobian() relating to order of elements in pd vector.
Fortran uses column-major ordering to flatten a 2D matrix onto a 1D array.
However, the way I was doing it in dassl.jacobian() was row-major ordering
(which is used by C). This was causing dassl.jacobian() to inadvertently
and silently return the transpose of the Jacobian matrix.

---
I hope this bug fix applies the same way to this branch
(I copied this commit from Josh's branch which has diverged a bit)
Signed-off-by: Richard West <r.west@neu.edu>